+++ packages/wiz-ui-next/src/components/base/inputs/search-selector/popup-button-group/popup-button-group.vue
@@ -49,6 +49,7 @@
             paddingLeft: `calc(${THEME.spacing.xs} + ${depth} * ${THEME.spacing.lg})`,
           }"
           :tabIndex="0"
+          :aria-label="`search-selector-${item.item.option.label}`"
           @click="popupButtonMouseDown(item.item)"
           @mouseover="popupButtonMouseOver(item.item)"
           @mouseout="popupButtonMouseOut(item.item)"