+++ packages/wiz-ui-next/src/components/base/tag/tag.vue
@@ -60,8 +60,10 @@ defineProps({
     type: String as PropType<
       | "info"
       | "red"
+      | "redFill"
       | "green"
       | "yellow"
+      | "yellowFill"
       | "blue"
       | "mono"
       | "gray"