+++ packages/wiz-ui-react/src/components/custom/notification/components/notification-list.tsx
@@ -1,13 +1,12 @@
 import { ComponentName } from "@wizleap-inc/wiz-ui-constants";
-import { FC, Fragment } from "react";
+import { FC, Fragment, ComponentPropsWithoutRef } from "react";
 
 import { WizBox, WizDivider, WizHStack, WizText } from "@/components";
-import { BaseProps } from "@/types";
 
 import { WizNotificationPanel } from "./notification-panel";
 import { NotificationItem, PanelVariant } from "./types";
 
-type Props = BaseProps & {
+type Props = Omit<ComponentPropsWithoutRef<"div">, "onClick"> & {
   variant?: PanelVariant;
   notifications: NotificationItem[];
   height?: string;
@@ -16,21 +15,15 @@ type Props = BaseProps & {
 };
 
 const NotificationList: FC<Props> = ({
-  className,
-  style,
   variant,
   notifications,
   height,
   emptyMessage,
   onClick,
+  ...props
 }) => {
   return (
-    <WizBox
-      className={className}
-      style={style}
-      height={height}
-      overflowY="scroll"
-    >
+    <WizBox {...props} height={height} overflowY="scroll">
       {notifications.map((notification, i) => (
         <Fragment key={notification.id}>
           <WizNotificationPanel